--- .idea/libraries/Maven__junit_junit_4_13_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: junit:junit:4.13.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.13.2/junit-4.13.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.13.2/junit-4.13.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.13.2/junit-4.13.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
