--- .idea/libraries/Maven__net_bytebuddy_byte_buddy_1_11_16.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.bytebuddy:byte-buddy:1.11.16">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.11.16/byte-buddy-1.11.16.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.11.16/byte-buddy-1.11.16-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.11.16/byte-buddy-1.11.16-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
