--- .idea/libraries/Maven__org_mockito_mockito_core_3_12_4.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.mockito:mockito-core:3.12.4">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mockito/mockito-core/3.12.4/mockito-core-3.12.4.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mockito/mockito-core/3.12.4/mockito-core-3.12.4-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/mockito/mockito-core/3.12.4/mockito-core-3.12.4-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
