--- .idea/libraries/Maven__org_objenesis_objenesis_3_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.objenesis:objenesis:3.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/objenesis/objenesis/3.2/objenesis-3.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/objenesis/objenesis/3.2/objenesis-3.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/objenesis/objenesis/3.2/objenesis-3.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
