+++ src/test/java/AnimalTest.java
@@ -1,40 +1,8 @@
 import com.example.Animal;
 import org.junit.Assert;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import java.util.List;
-
-@RunWith(Parameterized.class)
 public class AnimalTest {
-
-    @Rule
-    public ExpectedException thrown = ExpectedException.none();
-
-    private final String animalType;
-    private final List<String> expectedListOfFood;
-
-    public AnimalTest(String animalType, List<String> expectedListOfFood) {
-        this.animalType = animalType;
-        this.expectedListOfFood = expectedListOfFood;
-    }
-
-    @Parameterized.Parameters
-    public static Object[] getTestData() {
-        return new Object[][]{
-                {"Травоядное", List.of("Трава", "Различные растения")},
-                {"Хищник", List.of("Животные", "Птицы", "Рыба")},
-        };
-    }
-
-    @Test
-    public void getFoodTest() throws Exception {
-        Animal objAnimal = new Animal();
-        Assert.assertEquals("Ожидаемый результат не совпадает с фактическим", expectedListOfFood, objAnimal.getFood(animalType));
-    }
 
     // проверка что вызов getFood() у неизвестного животного выкинет нужное исключение
     @Test
@@ -42,7 +10,7 @@ public class AnimalTest {
         Animal unknownAnimal = new Animal();
         try {
             unknownAnimal.getFood("UnknownAnimal");
-//            Assert.fail("Expected Exception");
+            Assert.fail("Expected Exception");
         } catch (Exception thrown) {
             Assert.assertEquals("Неизвестный вид животного, используйте значение Травоядное или Хищник", thrown.getMessage());
         }
