+++ src/test/java/LionHasManeTest.java
@@ -32,7 +32,7 @@ public class LionHasManeTest {
 
     @Before
     public void init() {
-        MockitoAnnotations.initMocks(this);
+        MockitoAnnotations.openMocks(this);
     }
 
     @Test
